requirements.txt
appdirs==1.4.0
coreapi==2.2.4
coreschema==0.0.4
Django==1.10.5
django-auditlog==0.4.3
django-cors-headers==2.0.2
django-filter==1.0.1
django-jsonfield==1.0.1
django-rest-swagger==2.1.1
djangorestframework==3.5.4
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
openapi-codec==1.3.1
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
